OWA: Move more OWA JSON requests into separate files
