+++ app/logic/Mail/OWA/OWAAccount.ts
@@ -10,6 +10,7 @@ import type { OWAAddressbook } from "../../Contacts/OWA/OWAAddressbook";
 import { newCalendarForProtocol} from "../../Calendar/AccountsList/Calendars";
 import type { OWACalendar } from "../../Calendar/OWA/OWACalendar";
 import OWACreateItemRequest from "./Request/OWACreateItemRequest";
+import OWASubscribeToNotificationRequest from "./Request/OWASubscribeToNotificationRequest";
 import { OWALoginBackground } from "./Login/OWALoginBackground";
 import { owaAutoFillLoginPage } from "./Login/OWALoginAutoFill";
 import type { PersonUID } from "../../Abstract/PersonUID";
@@ -110,8 +111,7 @@ export class OWAAccount extends MailAccount {
     calendar.account = this;
     await calendar.listEvents();
 
-    let request = new OWASubscribeToNotificationRequest();
-    await this.callOWA(request);
+    await this.callOWA(new OWASubscribeToNotificationRequest());
 
     this.notifications = this.isOffice365()
       ? new OWAOffice365Notifications(this)
@@ -347,7 +347,7 @@ export class OWAAccount extends MailAccount {
     }
   }
 
-  handleHierarchyNotification(notification: any) {
+  protected handleHierarchyNotification(notification: any) {
     try {
       let parent = this.folderMap.get(notification.parentFolderId);
       if (!parent && notification.parentFolderId != this.msgFolderRootID) {
@@ -384,33 +384,4 @@ function addRecipients(aRequest: any, aType: string, aRecipients: PersonUID[]): 
     Name: recipient.name,
     EmailAddress: recipient.emailAddress,
   })), "message:" + aType);
-}
-
-class OWASubscribeToNotificationRequest {
-  readonly request = {
-    __type: "NotificationSubscribeJsonRequest:#Exchange",
-    Header: {
-      __type: "JsonRequestHeaders:#Exchange",
-      RequestServerVersion: "Exchange2013",
-    },
-  };
-  readonly subscriptionData = [{
-    __type: "SubscriptionData:#Exchange",
-    SubscriptionId: "HierarchyNotification",
-    Parameters: {
-      __type: "SubscriptionParameters:#Exchange",
-      NotificationType: "HierarchyNotification",
-    },
-  }, {
-    __type: "SubscriptionData:#Exchange",
-    SubscriptionId: "NewMailNotification",
-    Parameters: {
-      __type: "SubscriptionParameters:#Exchange",
-      NotificationType: "NewMailNotification",
-    },
-  }];
-
-  get type() {
-    return "SubscribeToNotification";
-  }
 }
